Added a function for an arbitrary search term and sorted some things into alphabetical order
To aid a search component in the client, I have added another function in the server called getMealsByArbitraryString. This is basically the same as getMealByName - it even uses the same REST API endpoint. The difference is this returns an optional array of optional Meals. The point of getMealByName is to get *just one*, whereas this is a general purpose search function.